I think they're all back and most are working
