+++ src/BloomBrowserUI/publish/stories/accessibility.stories.tsx
@@ -1,0 +1,17 @@
+import * as React from "react";
+import type { Meta, StoryObj } from "@storybook/react";
+import { AccessibilityCheckScreen } from "../accessibilityCheck/accessibilityCheckScreen";
+import "../storiesApiMocks";
+
+const meta: Meta<typeof AccessibilityCheckScreen> = {
+    title: "Publish/Accessibility",
+    component: AccessibilityCheckScreen
+};
+
+export default meta;
+type Story = StoryObj<typeof AccessibilityCheckScreen>;
+
+export const AccessibilityCheckScreenStory: Story = {
+    name: "AccessibilityCheckScreen",
+    render: () => <AccessibilityCheckScreen />
+};
