+++ src/BloomBrowserUI/publish/stories/bloom-reader.stories.tsx
@@ -1,0 +1,17 @@
+import * as React from "react";
+import type { Meta, StoryObj } from "@storybook/react";
+import { ReaderPublishScreen } from "../ReaderPublish/ReaderPublishScreen";
+import "../storiesApiMocks";
+
+const meta: Meta<typeof ReaderPublishScreen> = {
+    title: "Publish/Bloom Reader",
+    component: ReaderPublishScreen
+};
+
+export default meta;
+type Story = StoryObj<typeof ReaderPublishScreen>;
+
+export const ReaderPublishScreenStory: Story = {
+    name: "ReaderPublishScreen",
+    render: () => <ReaderPublishScreen />
+};
