+++ src/BloomBrowserUI/publish/stories/epub.stories.tsx
@@ -1,0 +1,27 @@
+import * as React from "react";
+import type { Meta, StoryObj } from "@storybook/react";
+import { EPUBPublishScreen } from "../ePUBPublish/ePUBPublishScreen";
+import BookMetadataDialog from "../metadata/BookMetadataDialog";
+import "../storiesApiMocks";
+
+const meta: Meta = {
+    title: "Publish/ePUB"
+};
+
+export default meta;
+type Story = StoryObj;
+
+export const EpubPublishScreenStory: Story = {
+    name: "EPUBPublishScreen",
+    render: () => <EPUBPublishScreen />
+};
+
+export const BookMetadataDialogStory: Story = {
+    name: "BookMetadataDialog",
+    render: () => (
+        <BookMetadataDialog
+            startOpen={true}
+            onClose={() => alert("BookMetadataDialog closed with OK")}
+        />
+    )
+};
