+++ src/BloomBrowserUI/publish/stories/video.stories.tsx
@@ -1,0 +1,17 @@
+import * as React from "react";
+import type { Meta, StoryObj } from "@storybook/react";
+import { PublishAudioVideo } from "../video/PublishAudioVideo";
+import "../storiesApiMocks";
+
+const meta: Meta<typeof PublishAudioVideo> = {
+    title: "Publish/Video",
+    component: PublishAudioVideo
+};
+
+export default meta;
+type Story = StoryObj<typeof PublishAudioVideo>;
+
+export const PublishAudioVideoStory: Story = {
+    name: "PublishAudioVideo",
+    render: () => <PublishAudioVideo />
+};
